requirements.txt
google-adk>=0.1.0
python-dotenv>=1.0.0
loguru>=0.7.0
google-generativeai>=0.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
streamlit>=1.28.0
requests>=2.31.0
